discordBot/requirements.txt
requests
discord.py
beautifulsoup4